# lib/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: jambonz REST API
  description: jambonz REST API
  contact:
    email: daveh@drachtio.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: /v1
    description: development server
paths:
...
  /Users/{UserSid}:
    parameters:
      - name: UserSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
    put:
      summary: update user information
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email: 
                  type: string
                email_activation_code:
                  type: string
                old_password:
                  type: string
                  description: existing password, which is to be replaced
                new_password:
                  type: string
                  description: new password
      responses:
        204:
          description: user updated
        403:
          description: password change failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /Users/me:
    get:
      summary: retrieve details about logged-in user and associated account
      operationId: getMyDetails
      responses:
        200:
          description: full-ish detail about the logged-in user and account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAndAccountDetail'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
  /VoipCarriers:
    post:
      summary: create voip carrier
      operationId: createVoipCarrier
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: voip carrier name
                  example: fastco
                description:
                  type: string
                  example: my US sip trunking provider
                account_sid:
                  type: string
                application_sid:
                  type: string
                e164_leading_plus:
                  type: boolean
                  description: whether a leading + is required on INVITEs to this provider
                  example: true
                requires_register:
                  type: boolean
                  description: wehther this provider requires us to send a REGISTER to them in order to receive calls
                register_username:
                  type: string
                  description: sip username to authenticate with, if registration is required
                  example: foo
                register_sip_realm:
                  type: string
                  description: sip realm to authenticate with, if registration is required
                  example: sip.fastco.com
                register_password:
                  type: string
                  description: sip password to authenticate with, if registration is required
                  example: bar
                tech_prefix:
                  type: string
                  description: prefix to be applied to the called number for outbound call attempts
                inbound_auth_username:
                  type: string
                  description: challenge inbound calls with this username/password if supplied
                inbound_auth_password:
                  type: string
                  description: challenge inbound calls with this username/password if supplied
                diversion:
                  type: string
                  description: Diversion header or phone number to apply to outbound calls
              required: 
                - name
      responses:
        201:
          description: voip carrier successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulAdd'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      summary: list voip carriers
      operationId: listVoipCarriers
      responses:
        200:
          description: list of voip carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoipCarrier'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /VoipCarriers/{VoipCarrierSid}:
    parameters:
      - name: VoipCarrierSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
    delete:
      summary: delete a voip carrier
      operationId: deleteVoipCarrier
      responses:
        204:
          description: voip carrier successfully deleted
        404:
          description: voip carrier not found
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
                example:
                  msg: a service provider with active accounts can not be deleted
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      summary: retrieve voip carrier
      operationId: getVoipCarrier
      responses:
        200:
          description: voip carrier found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoipCarrier'
        404:
          description: voip carrier not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    put:
      summary: update voip carrier
      operationId: updateVoipCarrier
      parameters:
      - name: VoipCarrierSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoipCarrier'
      responses:
        204:
          description: voip carrier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoipCarrier'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        404:
          description: voip carrier not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
...
    SuccessfulAdd:
      type: object
      required:
        - sid
      properties:
        sid:
          type: string
      example:
        sid: 9d26a637-1679-471f-8da8-7150266e1254
    GeneralError:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
      example:
        msg: specific error detail will be provided here
...
    VoipCarrier:
      type: object
      properties:
        voip_carrier_sid:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        account_sid:
          type: string
          format: uuid
        application_sid:
          type: string
          format: uuid          
        e164_leading_plus:
          type: boolean
        requires_register:
          type: boolean
        register_username:
          type: string
        register_sip_realm:
          type: string
        register_password:
          type: string
        tech_prefix:
          type: string
        inbound_auth_username:
          type: string
        inbound_auth_password:
          type: string
        diversion:
          type: string
        is_active:
          type: boolean
      required:
        - voip_carrier_sid
        - name
...
    UserAndAccountDetail:
      type: object
      properties:
        user:
          type: object
          properties:
            user_sid:
              type: string
            name: 
              type: string
            email:
              type: string
            phone:
              type: string
            service_provider_sid:
              type: string
            force_change:
              type: boolean
            provider:
              type: string
            provider_userid:
              type: string
            scope:
              type: string
            email_validated:
              type: boolean
            phone_validated:
              type: string
        account:
          type: object
          properties:
            account_sid:
              type: string
            sip_realm:
              type: string
            service_provider_sid:
              type: string
            registration_hook_sid:
              type: string
            device_calling_application_sid:
              type: string
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
        testapp:
          type: object
          properties:
            application_sid:
              type: string
            name:
              type: string
            service_provider_sid:
              type: string
            account_sid:
              type: string
            call_hook_sid:
              type: string
            call_status_hook_sid:
              type: string
            messaging_hook_sid:
              type: string
            speech_synthesis_vendor:
              type: string
            speech_synthesis_language:
              type: string
            speech_synthesis_voice:
              type: string
            speech_recognizer_vendor:
              type: string
            speech_recognizer_language:
              type: string
        balance:
          type: object
          properties:
            currency:
              type: string
              enum:
                - USD 
                - EUR
            balance:
              type: number
            last_updated_at:
              type: string
              format: date-time
            created_at:
              type: string
              format: date-time
            last_transaction_id:
              type: string
        capacities:
          type: object
          properties:
            effective_start_date:
              type: string
              format: date-time
            effective_end_date:
              type: string
              format: date-time
            limit_sessions:
              type: integer
            limit_registrations:
              type: integer
        api_keys:
          type: array
          items:
            type: object
            properties:
              token:
                type: string
              last_used:
                type: string
                format: date-time
              created_at:
                type: string
                format: date-time
        products:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              category:
                type: string
              quantity:
                type: string
              in_starter_set:
                type: boolean
              product_sid:
                type: string
              effective_start_date:
                type: string
                format: date-time
              effective_end_date:
                type: string
                format: date-time
...
security:
  - bearerAuth: []